+++ Example/iOS Example/AppDelegate.swift
@@ -19,12 +19,12 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     func applicationDidEnterBackground(_ application: UIApplication) {
         // Use this method to release shared resources, save user data, invalidate timers, and store enough application state information to restore your application to its current state in case it is terminated later.
         // If your application supports background execution, this method is called instead of applicationWillTerminate: when the user quits.
-        Tracker.shared?.dispatch()
+        PiwikTracker.shared?.dispatch()
     }
     
     func applicationWillEnterForeground(_ application: UIApplication) {
         // Called as part of the transition from the background to the active state; here you can undo many of the changes made on entering the background.
-        Tracker.shared?.startNewSession()
+        PiwikTracker.shared?.startNewSession()
     }
     
     func applicationDidBecomeActive(_ application: UIApplication) {
@@ -42,7 +42,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
                 alert.show()
                 return
         }
-        Tracker.configureSharedInstance(withSiteID: siteid, baseURL: baseurl)
+        PiwikTracker.configureSharedInstance(withSiteID: siteid, baseURL: baseurl)
     }
     
 }