+++ Example/iOS Example/ObjectiveCCompatibilityChecker.m
@@ -5,10 +5,10 @@
 @implementation ObjectiveCCompatibilityChecker
 
 - (void)check {
-    [Tracker configureSharedInstanceWithSiteID:@"5" baseURL:[NSURL URLWithString:@"http://example.com/piwik.php"]];
-    [[Tracker shared] trackWithView:@[@"example"] url:nil];
-    [[Tracker shared] trackWithEventWithCategory:@"category" action:@"action" name:nil number:nil];
-    [[Tracker shared] dispatch];
+    [PiwikTracker configureSharedInstanceWithSiteID:@"5" baseURL:[NSURL URLWithString:@"http://example.com/piwik.php"]];
+    [[PiwikTracker shared] trackWithView:@[@"example"] url:nil];
+    [[PiwikTracker shared] trackWithEventWithCategory:@"category" action:@"action" name:nil number:nil];
+    [[PiwikTracker shared] dispatch];
 }
 
 @end